+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog
 
-* 2014-12-01: Initial v0.1 public release of the Welcomer Framework!!
+* 2014-12-03 v0.1.1: Fix some bugs that crept through..
+* 2014-12-01 v0.1.0: Initial v0.1 public release of the Welcomer Framework!!